Time: 5 ms (64.64%), Space: 15.7 MB (40.73%) - LeetHub

+++ 3195-separate-black-and-white-balls/3195-separate-black-and-white-balls.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long minimumSteps(string s) {
         int black=0;
-        long  swap=0;
+        long long swap=0;
         for(int i=0;i<s.length();i++){
 
             if(s[i]=='0'){
